Document the meaning of bitmap values
The actual meaning behind the pixels in generated bitmaps is unclear.
This ambiguity can lead to difficutly implemeniting shaders or using
msdfgen properly as a library.

Add a note to each function which generates a bitmap what the actual
meaning of the pixels is; I.E., what the range of values.